Get composite team talent rankings for all teams in a given year or all rankings from 2015 - 2018

Extracts team talent rankings as sourced from 247 rankings

cfb_team_talent(year = NULL)

Arguments

year
Year

Examples

cfb_team_talent()

#>     year                   school talent
#> 1   2018               Ohio State 984.30
#> 2   2018                  Alabama 978.54
#> 3   2018                  Georgia 964.00
#> 4   2018                      USC 933.65
#> 5   2018                  Clemson 893.21
#> 6   2018                      LSU 889.91
#> 7   2018            Florida State 888.75
#> 8   2018                 Michigan 862.35
#> 9   2018                    Texas 861.20
#> 10  2018               Notre Dame 847.85
#> 11  2018                 Oklahoma 843.79
#> 12  2018               Penn State 834.24
#> 13  2018                  Florida 831.78
#> 14  2018                   Auburn 827.37
#> 15  2018                Tennessee 826.98
#> 16  2018                Texas A&M 805.66
#> 17  2018                 Stanford 786.63
#> 18  2018               Washington 772.88
#> 19  2018                     UCLA 754.31
#> 20  2018                    Miami 753.72
#> 21  2018           South Carolina 749.98
#> 22  2018                 Ole Miss 734.00
#> 23  2018        Mississippi State 729.97
#> 24  2018                 Maryland 725.94
#> 25  2018                 Nebraska 711.22
#> 26  2018                   Oregon 695.81
#> 27  2018                 Arkansas 694.63
#> 28  2018           North Carolina 688.48
#> 29  2018            Arizona State 687.72
#> 30  2018            Virginia Tech 675.71
#> 31  2018                      TCU 671.88
#> 32  2018                 Kentucky 670.27
#> 33  2018                   Baylor 668.37
#> 34  2018           Michigan State 658.85
#> 35  2018                Wisconsin 653.50
#> 36  2018            West Virginia 646.91
#> 37  2018           Oklahoma State 645.01
#> 38  2018               Pittsburgh 637.96
#> 39  2018                     Duke 635.63
#> 40  2018               Louisville 632.22
#> 41  2018                 NC State 629.39
#> 42  2018                     Iowa 615.08
#> 43  2018               Vanderbilt 611.99
#> 44  2018             Northwestern 608.89
#> 45  2018               California 596.08
#> 46  2018                 Missouri 595.14
#> 47  2018                     Utah 594.60
#> 48  2018                Minnesota 588.97
#> 49  2018             Georgia Tech 583.49
#> 50  2018                  Arizona 580.54
#> 51  2018                  Houston 577.17
#> 52  2018               Iowa State 575.76
#> 53  2018                 Colorado 575.65
#> 54  2018                 Syracuse 570.20
#> 55  2018                      UCF 568.35
#> 56  2018                      BYU 565.85
#> 57  2018                  Indiana 564.04
#> 58  2018                 Illinois 562.67
#> 59  2018               Texas Tech 561.59
#> 60  2018                  Rutgers 555.24
#> 61  2018         Washington State 554.77
#> 62  2018            South Florida 548.62
#> 63  2018             Kansas State 547.43
#> 64  2018               Cincinnati 543.78
#> 65  2018              Boise State 543.43
#> 66  2018                 Virginia 542.40
#> 67  2018         Florida Atlantic 537.93
#> 68  2018             Oregon State 535.36
#> 69  2018           Boston College 533.43
#> 70  2018              Wake Forest 525.48
#> 71  2018                   Purdue 520.48
#> 72  2018                   Kansas 517.10
#> 73  2018                      SMU 485.46
#> 74  2018                  Memphis 480.07
#> 75  2018           Colorado State 479.97
#> 76  2018                   Temple 478.30
#> 77  2018           Louisiana Tech 477.50
#> 78  2018    Florida International 476.52
#> 79  2018          San Diego State 472.60
#> 80  2018                   Toledo 463.96
#> 81  2018             Fresno State 459.07
#> 82  2018                 Marshall 451.61
#> 83  2018                   Tulane 445.36
#> 84  2018         Western Michigan 442.29
#> 85  2018            East Carolina 434.48
#> 86  2018           Arkansas State 433.73
#> 87  2018                   Nevada 432.40
#> 88  2018        Northern Illinois 432.31
#> 89  2018           San José State 430.62
#> 90  2018               Miami (OH) 421.22
#> 91  2018         Middle Tennessee 414.19
#> 92  2018         Central Michigan 413.60
#> 93  2018               New Mexico 411.95
#> 94  2018           UT San Antonio 405.20
#> 95  2018            Georgia State 404.34
#> 96  2018                     UNLV 400.76
#> 97  2018            South Alabama 394.24
#> 98  2018              Connecticut 389.20
#> 99  2018                    UMass 388.91
#> 100 2018                      UAB 387.71
#> 101 2018     Southern Mississippi 387.32
#> 102 2018               Utah State 377.39
#> 103 2018                Louisiana 376.68
#> 104 2018            Bowling Green 375.14
#> 105 2018                     Rice 375.12
#> 106 2018                  Buffalo 369.14
#> 107 2018              North Texas 368.58
#> 108 2018                    Tulsa 368.14
#> 109 2018               Ball State 362.10
#> 110 2018                    Akron 361.00
#> 111 2018        Appalachian State 356.16
#> 112 2018         Western Kentucky 353.43
#> 113 2018                     Troy 352.36
#> 114 2018               Kent State 351.73
#> 115 2018                     Navy 349.26
#> 116 2018                  Wyoming 348.34
#> 117 2018                     Ohio 346.18
#> 118 2018                     UTEP 345.76
#> 119 2018         Eastern Michigan 344.06
#> 120 2018                  Hawai'i 343.97
#> 121 2018             Old Dominion 333.18
#> 122 2018         Louisiana Monroe 326.66
#> 123 2018                Charlotte 311.00
#> 124 2018         Georgia Southern 303.12
#> 125 2018              Texas State 302.55
#> 126 2018         Coastal Carolina 284.24
#> 127 2018         New Mexico State 249.32
#> 128 2018                     Army 230.75
#> 129 2018                Air Force 186.27
#> 130 2018       Jacksonville State 174.86
#> 131 2018                  Hampton 149.11
#> 132 2018            James Madison 147.11
#> 133 2018   Southeastern Louisiana 135.64
#> 134 2018        Sam Houston State 133.13
#> 135 2018        Sam Houston State 133.13
#> 136 2018              Austin Peay 123.40
#> 137 2018            Alabama State 118.68
#> 138 2018                 Duquesne 113.21
#> 139 2018              Chattanooga 112.15
#> 140 2018           Illinois State 109.44
#> 141 2018        Southern Illinois 108.59
#> 142 2018         Eastern Kentucky 107.58
#> 143 2018              Stony Brook 106.51
#> 144 2018            Montana State  98.98
#> 145 2018                UT Martin  98.34
#> 146 2018                  Liberty  93.99
#> 147 2018                    Idaho  87.35
#> 148 2018           Kennesaw State  86.45
#> 149 2018           Portland State  85.83
#> 150 2018         Northern Arizona  85.03
#> 151 2018                  Samford  84.29
#> 152 2018       North Dakota State  83.68
#> 153 2018                   Towson  83.58
#> 154 2018           Texas Southern  83.08
#> 155 2018          Tennessee State  82.27
#> 156 2018                 Richmond  82.20
#> 157 2018                 Delaware  79.23
#> 158 2018         Youngstown State  70.45
#> 159 2018                Grambling  69.99
#> 160 2018 Southeast Missouri State  69.85
#> 161 2018         Western Illinois  66.84
#> 162 2018                 Nicholls  66.08
#> 163 2018             Rhode Island  62.64
#> 164 2018          Bethune-Cookman  61.90
#> 165 2018                   Furman  59.57
#> 166 2018                 Campbell  58.46
#> 167 2018            Norfolk State  56.44
#> 168 2018           Missouri State  53.43
#> 169 2018              Weber State  51.74
#> 170 2018      Charleston Southern  51.01
#> 171 2018                  Montana  49.16
#> 172 2018             Prairie View  47.10
#> 173 2018        Stephen F. Austin  46.25
#> 174 2018                  McNeese  43.55
#> 175 2018        Abilene Christian  43.10
#> 176 2018       North Carolina A&T  39.98
#> 177 2018              Florida A&M  39.57
#> 178 2018             Murray State  37.82
#> 179 2018                 Cal Poly  37.28
#> 180 2018         Western Carolina  35.54
#> 181 2018             Morgan State  35.38
#> 182 2018     East Tennessee State  35.09
#> 183 2018            Northern Iowa  34.70
#> 184 2018                   Wagner  32.77
#> 185 2018                   Butler  31.53
#> 186 2018                   Albany  31.53
#> 187 2018   North Carolina Central  31.00
#> 188 2018            North Alabama  30.02
#> 189 2018              The Citadel  29.01
#> 190 2018       South Dakota State  28.73
#> 191 2018           William & Mary  28.69
#> 192 2018                 Monmouth  28.13
#> 193 2018           Tennessee Tech  27.77
#> 194 2018                Princeton  27.29
#> 195 2018     South Carolina State  27.01
#> 196 2018          Houston Baptist  26.80
#> 197 2018                   Bryant  24.33
#> 198 2018             North Dakota  23.99
#> 199 2018            Indiana State  23.88
#> 200 2018         Central Arkansas  23.59
#> 201 2018           Incarnate Word  23.03
#> 202 2018             South Dakota  22.43
#> 203 2018                 Southern  22.25
#> 204 2018           Delaware State  21.32
#> 205 2018       Northwestern State  20.95
#> 206 2018              Idaho State  20.50
#> 207 2018         Sacramento State  20.45
#> 208 2018             Alcorn State  20.41
#> 209 2018       Eastern Washington  19.95
#> 210 2018             Sacred Heart  19.62
#> 211 2018                    Maine  18.93
#> 212 2018                   Howard  17.60
#> 213 2018             Jacksonville  17.49
#> 214 2018                  Fordham  15.33
#> 215 2018            Jackson State  14.46
#> 216 2018             Gardner-Webb  14.11
#> 217 2018                   Mercer  13.96
#> 218 2018                    Lamar  13.91
#> 219 2018                  Cornell  13.71
#> 220 2018        Northern Colorado  13.41
#> 221 2018                  Harvard  13.33
#> 222 2018                  Stetson  13.33
#> 223 2018          St Francis (PA)  13.00
#> 224 2018                      VMI  11.90
#> 225 2018            Southern Utah  11.84
#> 226 2018                    Drake  11.10
#> 227 2018               Georgetown  11.10
#> 228 2018            Robert Morris   9.24
#> 229 2018      Arkansas-Pine Bluff   8.98
#> 230 2018                   Dayton   8.88
#> 231 2018               Holy Cross   7.31
#> 232 2018      Central Connecticut   7.09
#> 233 2018         Eastern Illinois   6.75
#> 234 2018                   Lehigh   6.66
#> 235 2018                     Elon   6.66
#> 236 2018                   Marist   4.44
#> 237 2018                     Yale   3.30
#> 238 2017                  Alabama 997.57
#> 239 2017               Ohio State 959.12
#> 240 2017                      USC 934.00
#> 241 2017                  Georgia 930.34
#> 242 2017            Florida State 924.97
#> 243 2017                      LSU 907.10
#> 244 2017                 Michigan 874.28
#> 245 2017                   Auburn 867.43
#> 246 2017                  Clemson 853.65
#> 247 2017               Notre Dame 837.71
#> 248 2017                     UCLA 830.58
#> 249 2017                Tennessee 817.34
#> 250 2017                    Texas 811.03
#> 251 2017                 Stanford 810.31
#> 252 2017                Texas A&M 807.27
#> 253 2017                 Oklahoma 801.67
#> 254 2017                 Ole Miss 795.70
#> 255 2017                  Florida 793.83
#> 256 2017               Penn State 779.26
#> 257 2017                    Miami 749.08
#> 258 2017            Arizona State 720.44
#> 259 2017                   Oregon 720.23
#> 260 2017                 Arkansas 717.01
#> 261 2017               Washington 706.52
#> 262 2017           South Carolina 702.52
#> 263 2017        Mississippi State 701.06
#> 264 2017           North Carolina 697.54
#> 265 2017                 Maryland 696.21
#> 266 2017                 Nebraska 693.46
#> 267 2017           Michigan State 680.94
#> 268 2017                      TCU 672.81
#> 269 2017                 Kentucky 671.45
#> 270 2017            Virginia Tech 662.18
#> 271 2017               Pittsburgh 660.36
#> 272 2017                   Baylor 657.22
#> 273 2017               Louisville 640.98
#> 274 2017                Wisconsin 639.76
#> 275 2017           Oklahoma State 633.57
#> 276 2017                     Utah 630.95
#> 277 2017                     Duke 627.62
#> 278 2017             Northwestern 627.54
#> 279 2017                 Missouri 623.33
#> 280 2017            West Virginia 613.18
#> 281 2017                 NC State 609.76
#> 282 2017               Vanderbilt 607.68
#> 283 2017                  Arizona 607.31
#> 284 2017                 Virginia 603.88
#> 285 2017               California 603.06
#> 286 2017                     Iowa 592.95
#> 287 2017               Texas Tech 574.93
#> 288 2017                  Rutgers 574.78
#> 289 2017             Georgia Tech 574.35
#> 290 2017                      BYU 574.09
#> 291 2017                      UCF 571.52
#> 292 2017                 Colorado 568.59
#> 293 2017               Iowa State 559.34
#> 294 2017                  Houston 558.72
#> 295 2017             Oregon State 557.02
#> 296 2017            South Florida 554.17
#> 297 2017         Washington State 549.42
#> 298 2017                  Indiana 547.92
#> 299 2017                Minnesota 543.80
#> 300 2017                 Syracuse 539.60
#> 301 2017                 Illinois 538.83
#> 302 2017           Boston College 535.21
#> 303 2017              Wake Forest 527.21
#> 304 2017             Kansas State 517.85
#> 305 2017                   Kansas 512.24
#> 306 2017                  Memphis 504.49
#> 307 2017                   Purdue 502.71
#> 308 2017              Boise State 501.53
#> 309 2017               Cincinnati 500.95
#> 310 2017                      SMU 492.63
#> 311 2017            East Carolina 471.10
#> 312 2017         Florida Atlantic 470.88
#> 313 2017                   Temple 468.29
#> 314 2017                 Marshall 445.48
#> 315 2017           Colorado State 441.98
#> 316 2017           Arkansas State 436.05
#> 317 2017                    Tulsa 422.77
#> 318 2017              Connecticut 420.17
#> 319 2017          San Diego State 419.43
#> 320 2017         Western Michigan 414.35
#> 321 2017               Miami (OH) 413.24
#> 322 2017                   Tulane 396.43
#> 323 2017           Louisiana Tech 394.31
#> 324 2017     Southern Mississippi 394.24
#> 325 2017            South Alabama 387.88
#> 326 2017         Central Michigan 387.20
#> 327 2017           San José State 378.56
#> 328 2017                     Navy 376.20
#> 329 2017             Fresno State 374.63
#> 330 2017        Northern Illinois 373.99
#> 331 2017                    UMass 367.59
#> 332 2017                   Toledo 366.47
#> 333 2017         Georgia Southern 365.57
#> 334 2017            Georgia State 364.16
#> 335 2017            Bowling Green 361.82
#> 336 2017                     Troy 352.10
#> 337 2017                     Rice 345.79
#> 338 2017         Middle Tennessee 344.65
#> 339 2017               New Mexico 344.64
#> 340 2017                   Nevada 341.11
#> 341 2017                    Akron 337.11
#> 342 2017        Appalachian State 335.87
#> 343 2017    Florida International 334.18
#> 344 2017         Western Kentucky 332.17
#> 345 2017             Old Dominion 331.20
#> 346 2017                     UNLV 331.18
#> 347 2017                  Hawai'i 330.56
#> 348 2017               Utah State 330.20
#> 349 2017                Louisiana 329.24
#> 350 2017                     Ohio 326.14
#> 351 2017               Ball State 318.17
#> 352 2017         Louisiana Monroe 303.96
#> 353 2017              North Texas 303.13
#> 354 2017                  Buffalo 294.08
#> 355 2017         Eastern Michigan 292.11
#> 356 2017           UT San Antonio 291.62
#> 357 2017               Kent State 288.56
#> 358 2017                Charlotte 283.01
#> 359 2017                  Wyoming 277.07
#> 360 2017              Texas State 274.67
#> 361 2017                     UTEP 264.42
#> 362 2017         New Mexico State 233.47
#> 363 2017                    Idaho 184.75
#> 364 2017            James Madison 179.50
#> 365 2017                      UAB 170.05
#> 366 2017           Portland State 147.63
#> 367 2017         Coastal Carolina 119.43
#> 368 2017              Stony Brook 111.94
#> 369 2017                   Towson 106.93
#> 370 2017                     Army  88.72
#> 371 2017                  Montana  80.03
#> 372 2017                 Cal Poly  76.98
#> 373 2017      Charleston Southern  72.38
#> 374 2017                 Richmond  71.64
#> 375 2017           Kennesaw State  70.97
#> 376 2017         Northern Arizona  69.20
#> 377 2017             Rhode Island  65.72
#> 378 2017              Weber State  51.74
#> 379 2017         Sacramento State  46.67
#> 380 2017           William & Mary  40.05
#> 381 2017                    Maine  38.99
#> 382 2017            Southern Utah  33.11
#> 383 2017        Northern Colorado  32.58
#> 384 2017            Montana State  26.40
#> 385 2017             Gardner-Webb  23.01
#> 386 2017                 Monmouth  21.44
#> 387 2017                Air Force  21.10
#> 388 2017       Eastern Washington  19.95
#> 389 2017                  Liberty  19.85
#> 390 2017              Idaho State  15.22
#> 391 2017     Presbyterian College  14.91
#> 392 2017                     Elon  13.41
#> 393 2017                Villanova  11.65
#> 394 2017            New Hampshire  11.21
#> 395 2016                  Alabama 982.66
#> 396 2016                      USC 936.86
#> 397 2016                      LSU 917.72
#> 398 2016            Florida State 906.64
#> 399 2016               Ohio State 902.46
#> 400 2016                  Georgia 872.00
#> 401 2016                   Auburn 865.36
#> 402 2016                 Michigan 852.78
#> 403 2016                  Clemson 846.00
#> 404 2016               Notre Dame 842.35
#> 405 2016                    Texas 838.97
#> 406 2016                     UCLA 819.16
#> 407 2016                Texas A&M 814.63
#> 408 2016                Tennessee 813.92
#> 409 2016                 Ole Miss 808.15
#> 410 2016                  Florida 798.44
#> 411 2016                 Stanford 778.45
#> 412 2016                 Oklahoma 777.48
#> 413 2016                    Miami 745.93
#> 414 2016               Penn State 742.52
#> 415 2016                   Oregon 738.29
#> 416 2016           Michigan State 719.70
#> 417 2016           South Carolina 706.90
#> 418 2016               Washington 699.64
#> 419 2016            Arizona State 694.21
#> 420 2016                 Nebraska 691.61
#> 421 2016                 Arkansas 689.41
#> 422 2016        Mississippi State 679.14
#> 423 2016           North Carolina 669.61
#> 424 2016                      TCU 665.53
#> 425 2016            West Virginia 654.64
#> 426 2016               Louisville 654.13
#> 427 2016                 Kentucky 652.07
#> 428 2016                 Maryland 651.99
#> 429 2016               Pittsburgh 649.85
#> 430 2016            Virginia Tech 646.82
#> 431 2016                Wisconsin 643.76
#> 432 2016                 Missouri 633.76
#> 433 2016           Oklahoma State 631.68
#> 434 2016                   Baylor 619.18
#> 435 2016               Vanderbilt 619.03
#> 436 2016                 Virginia 618.91
#> 437 2016                  Arizona 616.31
#> 438 2016                 NC State 616.29
#> 439 2016             Northwestern 613.55
#> 440 2016               Texas Tech 604.93
#> 441 2016               California 601.17
#> 442 2016                     Duke 599.82
#> 443 2016                  Rutgers 585.47
#> 444 2016             Georgia Tech 582.55
#> 445 2016                     Utah 580.07
#> 446 2016                  Houston 572.70
#> 447 2016                  Indiana 565.47
#> 448 2016            South Florida 563.99
#> 449 2016                     Iowa 558.73
#> 450 2016               Iowa State 552.17
#> 451 2016              Boise State 548.82
#> 452 2016               Cincinnati 541.58
#> 453 2016                Minnesota 535.44
#> 454 2016           Boston College 533.39
#> 455 2016                      BYU 529.05
#> 456 2016                 Colorado 526.39
#> 457 2016             Oregon State 524.89
#> 458 2016                 Illinois 523.91
#> 459 2016         Washington State 522.58
#> 460 2016                 Syracuse 520.26
#> 461 2016              Wake Forest 517.63
#> 462 2016                   Purdue 512.72
#> 463 2016                      UCF 510.27
#> 464 2016             Kansas State 505.29
#> 465 2016                   Temple 492.72
#> 466 2016                   Kansas 491.00
#> 467 2016                 Marshall 485.46
#> 468 2016          San Diego State 482.31
#> 469 2016                      SMU 474.03
#> 470 2016           Arkansas State 471.45
#> 471 2016                   Toledo 460.86
#> 472 2016           Louisiana Tech 451.38
#> 473 2016                  Memphis 451.24
#> 474 2016                    Akron 443.73
#> 475 2016         Western Michigan 443.16
#> 476 2016            East Carolina 442.26
#> 477 2016     Southern Mississippi 441.29
#> 478 2016           San José State 437.36
#> 479 2016         Middle Tennessee 436.60
#> 480 2016         Georgia Southern 435.33
#> 481 2016              Connecticut 423.56
#> 482 2016                Louisiana 421.37
#> 483 2016               Miami (OH) 418.04
#> 484 2016           Colorado State 412.28
#> 485 2016                    Tulsa 409.17
#> 486 2016         Florida Atlantic 408.40
#> 487 2016             Fresno State 404.97
#> 488 2016        Northern Illinois 403.27
#> 489 2016         Western Kentucky 401.82
#> 490 2016           UT San Antonio 397.64
#> 491 2016                     Rice 394.83
#> 492 2016            Bowling Green 392.66
#> 493 2016         Central Michigan 391.81
#> 494 2016                     Troy 390.47
#> 495 2016                   Nevada 389.56
#> 496 2016            South Alabama 389.36
#> 497 2016                   Tulane 388.04
#> 498 2016               New Mexico 387.26
#> 499 2016                  Hawai'i 384.97
#> 500 2016                     UNLV 376.26
#> 501 2016                     Ohio 372.56
#> 502 2016                    UMass 370.90
#> 503 2016               Ball State 365.94
#> 504 2016    Florida International 362.84
#> 505 2016               Utah State 357.13
#> 506 2016              Texas State 351.76
#> 507 2016              North Texas 351.58
#> 508 2016         Eastern Michigan 347.70
#> 509 2016             Old Dominion 347.20
#> 510 2016                  Buffalo 343.64
#> 511 2016                     Navy 342.77
#> 512 2016               Kent State 334.01
#> 513 2016        Appalachian State 333.36
#> 514 2016                Charlotte 330.55
#> 515 2016         Louisiana Monroe 322.82
#> 516 2016            Georgia State 321.79
#> 517 2016                  Wyoming 316.18
#> 518 2016                     UTEP 290.40
#> 519 2016         New Mexico State 267.66
#> 520 2016                    Idaho 256.78
#> 521 2016       Jacksonville State 254.26
#> 522 2016         Eastern Kentucky 184.88
#> 523 2016                     Yale 165.17
#> 524 2016            Northern Iowa 164.16
#> 525 2016           Portland State 154.68
#> 526 2016                     Army 131.74
#> 527 2016         Youngstown State 130.84
#> 528 2016            James Madison 128.58
#> 529 2016        Southern Illinois 123.69
#> 530 2016   Southeastern Louisiana 121.61
#> 531 2016              Chattanooga 119.20
#> 532 2016         Eastern Illinois 119.04
#> 533 2016                   Wagner 118.43
#> 534 2016              Stony Brook 118.15
#> 535 2016                 Duquesne 116.15
#> 536 2016                UT Martin 107.98
#> 537 2016         Northern Arizona 104.38
#> 538 2016           Illinois State 102.21
#> 539 2016                  Hampton  99.46
#> 540 2016                   Towson  95.18
#> 541 2016                  Samford  94.57
#> 542 2016                Grambling  93.31
#> 543 2016                 Cal Poly  91.73
#> 544 2016                 Nicholls  89.59
#> 545 2016            Jackson State  86.85
#> 546 2016                  Montana  83.10
#> 547 2016       North Dakota State  81.65
#> 548 2016        Stephen F. Austin  80.33
#> 549 2016         Western Illinois  78.62
#> 550 2016 Southeast Missouri State  76.33
#> 551 2016     South Carolina State  75.71
#> 552 2016         Coastal Carolina  71.67
#> 553 2016                  McNeese  69.85
#> 554 2016            Southern Utah  68.79
#> 555 2016             Prairie View  67.68
#> 556 2016            Norfolk State  67.07
#> 557 2016                   Furman  66.86
#> 558 2016      Charleston Southern  65.60
#> 559 2016            Montana State  64.33
#> 560 2016          Bethune-Cookman  63.39
#> 561 2016             Murray State  62.79
#> 562 2016         Central Arkansas  62.11
#> 563 2016          Houston Baptist  61.17
#> 564 2016             Pennsylvania  60.95
#> 565 2016                   Mercer  57.71
#> 566 2016           Kennesaw State  54.76
#> 567 2016              Weber State  54.68
#> 568 2016            Alabama State  53.46
#> 569 2016           Tennessee Tech  53.00
#> 570 2016                 Richmond  51.95
#> 571 2016          Tennessee State  51.82
#> 572 2016                 Southern  46.97
#> 573 2016         Sacramento State  45.18
#> 574 2016       Eastern Washington  45.10
#> 575 2016              Austin Peay  44.67
#> 576 2016                Air Force  44.58
#> 577 2016           Texas Southern  43.86
#> 578 2016                    Lamar  43.84
#> 579 2016            Indiana State  43.84
#> 580 2016           William & Mary  43.54
#> 581 2016     Presbyterian College  41.40
#> 582 2016             Rhode Island  41.24
#> 583 2016                 Campbell  40.68
#> 584 2016             Alcorn State  39.30
#> 585 2016              The Citadel  37.64
#> 586 2016                  Fordham  36.98
#> 587 2016                Princeton  36.73
#> 588 2016                 Monmouth  35.77
#> 589 2016             South Dakota  32.89
#> 590 2016       Northwestern State  32.06
#> 591 2016                      VMI  31.43
#> 592 2016     East Tennessee State  31.35
#> 593 2016               Valparaiso  29.84
#> 594 2016        Northern Colorado  29.66
#> 595 2016                    Drake  27.37
#> 596 2016            Robert Morris  26.70
#> 597 2016              Alabama A&M  26.31
#> 598 2016              Idaho State  25.18
#> 599 2016                Villanova  23.98
#> 600 2016           Savannah State  23.51
#> 601 2016                    Maine  23.37
#> 602 2016           Delaware State  22.84
#> 603 2016             Gardner-Webb  22.66
#> 604 2016                   Butler  21.56
#> 605 2016                 Columbia  21.24
#> 606 2016                  Liberty  21.04
#> 607 2016        Abilene Christian  19.77
#> 608 2016                  Wofford  19.48
#> 609 2016             Sacred Heart  19.44
#> 610 2016                  Harvard  17.28
#> 611 2016                    Brown  16.92
#> 612 2016              Florida A&M  16.64
#> 613 2016                  Cornell  15.89
#> 614 2016                   Dayton  14.99
#> 615 2016               Holy Cross  14.75
#> 616 2016         Western Carolina  13.67
#> 617 2016             Jacksonville  13.41
#> 618 2016   North Carolina Central  13.15
#> 619 2016           Missouri State  12.70
#> 620 2016               Georgetown  12.60
#> 621 2016            New Hampshire  11.21
#> 622 2016                  Stetson   9.99
#> 623 2016                   Howard   7.53
#> 624 2016      Central Connecticut   6.66
#> 625 2016                   Lehigh   6.66
#> 626 2016      Arkansas-Pine Bluff   6.66
#> 627 2016           Morehead State   6.66
#> 628 2016                San Diego   6.66
#> 629 2016                   Bryant   6.24
#> 630 2016       North Carolina A&T   3.81
#> 631 2016          St Francis (PA)   1.74
#> 632 2015                  Alabama 981.90
#> 633 2015                      USC 926.71
#> 634 2015               Ohio State 907.32
#> 635 2015            Florida State 889.92
#> 636 2015                      LSU 889.37
#> 637 2015                  Georgia 877.74
#> 638 2015                   Auburn 865.41
#> 639 2015               Notre Dame 857.33
#> 640 2015                 Michigan 851.12
#> 641 2015                Texas A&M 830.34
#> 642 2015                    Texas 821.26
#> 643 2015                Tennessee 810.53
#> 644 2015                  Clemson 807.56
#> 645 2015                     UCLA 793.22
#> 646 2015                  Florida 789.79
#> 647 2015                 Oklahoma 775.90
#> 648 2015                    Miami 773.23
#> 649 2015                 Stanford 765.98
#> 650 2015                 Ole Miss 764.34
#> 651 2015                   Oregon 757.52
#> 652 2015               Penn State 732.36
#> 653 2015           South Carolina 725.72
#> 654 2015           Michigan State 715.30
#> 655 2015                 Nebraska 699.89
#> 656 2015                 Arkansas 695.04
#> 657 2015           North Carolina 683.03
#> 658 2015            Arizona State 680.87
#> 659 2015        Mississippi State 680.06
#> 660 2015               Washington 667.45
#> 661 2015                 Virginia 667.05
#> 662 2015                   Baylor 660.35
#> 663 2015            Virginia Tech 658.38
#> 664 2015           Oklahoma State 651.83
#> 665 2015                 Missouri 642.93
#> 666 2015               Texas Tech 641.07
#> 667 2015               Louisville 641.05
#> 668 2015                Wisconsin 635.77
#> 669 2015                 Kentucky 634.52
#> 670 2015                      TCU 630.38
#> 671 2015               California 626.00
#> 672 2015            West Virginia 625.85
#> 673 2015               Pittsburgh 611.18
#> 674 2015                 NC State 610.33
#> 675 2015               Vanderbilt 609.76
#> 676 2015                  Arizona 606.92
#> 677 2015                 Maryland 603.76
#> 678 2015                  Rutgers 602.35
#> 679 2015             Northwestern 595.75
#> 680 2015             Georgia Tech 590.34
#> 681 2015                     Utah 574.41
#> 682 2015                     Iowa 573.06
#> 683 2015              Boise State 568.96
#> 684 2015                     Duke 563.19
#> 685 2015                  Indiana 559.08
#> 686 2015             Oregon State 544.66
#> 687 2015                 Colorado 542.06
#> 688 2015         Washington State 540.86
#> 689 2015           Boston College 534.60
#> 690 2015                   Purdue 534.26
#> 691 2015                 Illinois 530.96
#> 692 2015            South Florida 529.65
#> 693 2015               Iowa State 529.12
#> 694 2015                      BYU 521.43
#> 695 2015                Minnesota 516.10
#> 696 2015                  Houston 511.45
#> 697 2015                 Syracuse 508.80
#> 698 2015             Kansas State 503.64
#> 699 2015                   Kansas 499.97
#> 700 2015              Wake Forest 497.48
#> 701 2015          San Diego State 494.77
#> 702 2015               Cincinnati 494.65
#> 703 2015                 Marshall 489.77
#> 704 2015                      UCF 484.54
#> 705 2015     Southern Mississippi 470.84
#> 706 2015                      SMU 446.83
#> 707 2015         Western Michigan 442.13
#> 708 2015           Louisiana Tech 437.99
#> 709 2015            Bowling Green 433.59
#> 710 2015                Louisiana 430.17
#> 711 2015                   Toledo 429.03
#> 712 2015                    Akron 412.65
#> 713 2015         Middle Tennessee 409.82
#> 714 2015                   Tulane 408.22
#> 715 2015           Arkansas State 408.08
#> 716 2015                   Temple 405.92
#> 717 2015                    Tulsa 404.03
#> 718 2015            East Carolina 403.62
#> 719 2015             Fresno State 402.58
#> 720 2015              Connecticut 402.23
#> 721 2015         Florida Atlantic 401.61
#> 722 2015                     Rice 399.58
#> 723 2015               Miami (OH) 397.13
#> 724 2015         Georgia Southern 391.18
#> 725 2015           Colorado State 389.10
#> 726 2015                  Hawai'i 385.63
#> 727 2015                  Memphis 381.80
#> 728 2015              North Texas 380.69
#> 729 2015        Northern Illinois 380.69
#> 730 2015                   Nevada 380.53
#> 731 2015           San José State 379.60
#> 732 2015         Central Michigan 376.81
#> 733 2015            South Alabama 367.16
#> 734 2015              Texas State 360.49
#> 735 2015    Florida International 360.11
#> 736 2015           UT San Antonio 356.37
#> 737 2015         Western Kentucky 353.50
#> 738 2015               Ball State 349.94
#> 739 2015               New Mexico 341.72
#> 740 2015         Eastern Michigan 341.36
#> 741 2015               Kent State 338.40
#> 742 2015            Georgia State 335.67
#> 743 2015                     Troy 335.43
#> 744 2015                     UNLV 332.69
#> 745 2015               Utah State 331.44
#> 746 2015                     Ohio 330.61
#> 747 2015                    UMass 323.73
#> 748 2015             Old Dominion 319.60
#> 749 2015                  Buffalo 316.44
#> 750 2015                  Wyoming 314.72
#> 751 2015        Appalachian State 312.88
#> 752 2015                     Navy 306.87
#> 753 2015                Charlotte 304.05
#> 754 2015         Louisiana Monroe 292.73
#> 755 2015                    Idaho 262.39
#> 756 2015                     UTEP 256.12
#> 757 2015         New Mexico State 251.16
#> 758 2015                      UAB 236.34
#> 759 2015            Northern Iowa 202.77
#> 760 2015                     Army 186.45
#> 761 2015                     Yale 181.76
#> 762 2015       Jacksonville State 173.32
#> 763 2015              Chattanooga 137.06
#> 764 2015           Portland State 123.61
#> 765 2015         Youngstown State 116.79
#> 766 2015          Bethune-Cookman 116.09
#> 767 2015     South Carolina State 105.01
#> 768 2015        Stephen F. Austin 103.72
#> 769 2015                 Cal Poly  97.73
#> 770 2015                  Samford  95.68
#> 771 2015           Illinois State  92.42
#> 772 2015             Murray State  89.72
#> 773 2015            James Madison  84.74
#> 774 2015       North Dakota State  82.26
#> 775 2015            Jackson State  78.26
#> 776 2015            Indiana State  74.60
#> 777 2015          Tennessee State  74.37
#> 778 2015                   Towson  74.05
#> 779 2015             Pennsylvania  73.07
#> 780 2015                Air Force  70.95
#> 781 2015         Northern Arizona  68.35
#> 782 2015        Abilene Christian  66.43
#> 783 2015                  Montana  64.46
#> 784 2015                   Wagner  64.14
#> 785 2015                UT Martin  62.12
#> 786 2015              Stony Brook  61.96
#> 787 2015                   Furman  61.89
#> 788 2015        Southern Illinois  60.76
#> 789 2015                  Hampton  58.39
#> 790 2015   Southeastern Louisiana  58.08
#> 791 2015                Villanova  57.42
#> 792 2015               Valparaiso  56.37
#> 793 2015         Western Illinois  51.28
#> 794 2015                  McNeese  51.19
#> 795 2015                 Duquesne  50.43
#> 796 2015      Charleston Southern  48.43
#> 797 2015              The Citadel  47.12
#> 798 2015         Eastern Illinois  46.56
#> 799 2015         Eastern Kentucky  46.50
#> 800 2015                 Richmond  46.34
#> 801 2015       Eastern Washington  45.10
#> 802 2015           Texas Southern  43.10
#> 803 2015                Grambling  41.62
#> 804 2015         Coastal Carolina  41.50
#> 805 2015          Houston Baptist  39.71
#> 806 2015                 Southern  37.48
#> 807 2015         Sacramento State  37.46
#> 808 2015             Rhode Island  37.17
#> 809 2015              Austin Peay  36.99
#> 810 2015                  Fordham  36.98
#> 811 2015        Northern Colorado  36.95
#> 812 2015         Central Arkansas  35.29
#> 813 2015                   Mercer  35.02
#> 814 2015            Norfolk State  32.59
#> 815 2015           Tennessee Tech  31.74
#> 816 2015                      VMI  31.43
#> 817 2015                 Campbell  31.21
#> 818 2015                  Harvard  30.49
#> 819 2015 Southeast Missouri State  29.79
#> 820 2015                  Cornell  29.11
#> 821 2015       Northwestern State  28.02
#> 822 2015     Presbyterian College  25.63
#> 823 2015           Delaware State  25.26
#> 824 2015       South Dakota State  25.25
#> 825 2015                    Lamar  24.84
#> 826 2015                 Nicholls  24.71
#> 827 2015            Alabama State  24.46
#> 828 2015           Kennesaw State  23.49
#> 829 2015              Idaho State  21.69
#> 830 2015                 Columbia  21.24
#> 831 2015            New Hampshire  21.17
#> 832 2015             Prairie View  20.76
#> 833 2015     East Tennessee State  20.07
#> 834 2015              Weber State  19.80
#> 835 2015                  Liberty  19.57
#> 836 2015                  Wofford  19.48
#> 837 2015             Jacksonville  18.58
#> 838 2015                    Brown  16.92
#> 839 2015         Western Carolina  15.77
#> 840 2015                 Davidson  15.67
#> 841 2015           William & Mary  15.65
#> 842 2015              Florida A&M  15.60
#> 843 2015           Incarnate Word  15.59
#> 844 2015   North Carolina Central  14.26
#> 845 2015              Alabama A&M  13.33
#> 846 2015            Montana State  11.86
#> 847 2015                   Butler   9.96
#> 848 2015             Gardner-Webb   8.58
#> 849 2015                 Monmouth   7.93
#> 850 2015               Holy Cross   7.45
#> 851 2015                    Drake   7.43
#> 852 2015                    Maine   7.24
#> 853 2015                   Lehigh   6.66
#> 854 2015                  Stetson   6.66
#> 855 2015      Arkansas-Pine Bluff   6.66
#> 856 2015            Southern Utah   6.66
#> 857 2015                 UC Davis   6.66
#> 858 2015             South Dakota   3.36
#> 859 2015            Robert Morris   3.06
#> 860 2015                Dartmouth   2.96
#> 861 2015                   Marist   2.63
#> 862 2015                   Howard   2.22
#> 863 2015          St Francis (PA)   1.74

cfb_team_talent(year=2018)

#>     year                   school talent
#> 1   2018               Ohio State 984.30
#> 2   2018                  Alabama 978.54
#> 3   2018                  Georgia 964.00
#> 4   2018                      USC 933.65
#> 5   2018                  Clemson 893.21
#> 6   2018                      LSU 889.91
#> 7   2018            Florida State 888.75
#> 8   2018                 Michigan 862.35
#> 9   2018                    Texas 861.20
#> 10  2018               Notre Dame 847.85
#> 11  2018                 Oklahoma 843.79
#> 12  2018               Penn State 834.24
#> 13  2018                  Florida 831.78
#> 14  2018                   Auburn 827.37
#> 15  2018                Tennessee 826.98
#> 16  2018                Texas A&M 805.66
#> 17  2018                 Stanford 786.63
#> 18  2018               Washington 772.88
#> 19  2018                     UCLA 754.31
#> 20  2018                    Miami 753.72
#> 21  2018           South Carolina 749.98
#> 22  2018                 Ole Miss 734.00
#> 23  2018        Mississippi State 729.97
#> 24  2018                 Maryland 725.94
#> 25  2018                 Nebraska 711.22
#> 26  2018                   Oregon 695.81
#> 27  2018                 Arkansas 694.63
#> 28  2018           North Carolina 688.48
#> 29  2018            Arizona State 687.72
#> 30  2018            Virginia Tech 675.71
#> 31  2018                      TCU 671.88
#> 32  2018                 Kentucky 670.27
#> 33  2018                   Baylor 668.37
#> 34  2018           Michigan State 658.85
#> 35  2018                Wisconsin 653.50
#> 36  2018            West Virginia 646.91
#> 37  2018           Oklahoma State 645.01
#> 38  2018               Pittsburgh 637.96
#> 39  2018                     Duke 635.63
#> 40  2018               Louisville 632.22
#> 41  2018                 NC State 629.39
#> 42  2018                     Iowa 615.08
#> 43  2018               Vanderbilt 611.99
#> 44  2018             Northwestern 608.89
#> 45  2018               California 596.08
#> 46  2018                 Missouri 595.14
#> 47  2018                     Utah 594.60
#> 48  2018                Minnesota 588.97
#> 49  2018             Georgia Tech 583.49
#> 50  2018                  Arizona 580.54
#> 51  2018                  Houston 577.17
#> 52  2018               Iowa State 575.76
#> 53  2018                 Colorado 575.65
#> 54  2018                 Syracuse 570.20
#> 55  2018                      UCF 568.35
#> 56  2018                      BYU 565.85
#> 57  2018                  Indiana 564.04
#> 58  2018                 Illinois 562.67
#> 59  2018               Texas Tech 561.59
#> 60  2018                  Rutgers 555.24
#> 61  2018         Washington State 554.77
#> 62  2018            South Florida 548.62
#> 63  2018             Kansas State 547.43
#> 64  2018               Cincinnati 543.78
#> 65  2018              Boise State 543.43
#> 66  2018                 Virginia 542.40
#> 67  2018         Florida Atlantic 537.93
#> 68  2018             Oregon State 535.36
#> 69  2018           Boston College 533.43
#> 70  2018              Wake Forest 525.48
#> 71  2018                   Purdue 520.48
#> 72  2018                   Kansas 517.10
#> 73  2018                      SMU 485.46
#> 74  2018                  Memphis 480.07
#> 75  2018           Colorado State 479.97
#> 76  2018                   Temple 478.30
#> 77  2018           Louisiana Tech 477.50
#> 78  2018    Florida International 476.52
#> 79  2018          San Diego State 472.60
#> 80  2018                   Toledo 463.96
#> 81  2018             Fresno State 459.07
#> 82  2018                 Marshall 451.61
#> 83  2018                   Tulane 445.36
#> 84  2018         Western Michigan 442.29
#> 85  2018            East Carolina 434.48
#> 86  2018           Arkansas State 433.73
#> 87  2018                   Nevada 432.40
#> 88  2018        Northern Illinois 432.31
#> 89  2018           San José State 430.62
#> 90  2018               Miami (OH) 421.22
#> 91  2018         Middle Tennessee 414.19
#> 92  2018         Central Michigan 413.60
#> 93  2018               New Mexico 411.95
#> 94  2018           UT San Antonio 405.20
#> 95  2018            Georgia State 404.34
#> 96  2018                     UNLV 400.76
#> 97  2018            South Alabama 394.24
#> 98  2018              Connecticut 389.20
#> 99  2018                    UMass 388.91
#> 100 2018                      UAB 387.71
#> 101 2018     Southern Mississippi 387.32
#> 102 2018               Utah State 377.39
#> 103 2018                Louisiana 376.68
#> 104 2018            Bowling Green 375.14
#> 105 2018                     Rice 375.12
#> 106 2018                  Buffalo 369.14
#> 107 2018              North Texas 368.58
#> 108 2018                    Tulsa 368.14
#> 109 2018               Ball State 362.10
#> 110 2018                    Akron 361.00
#> 111 2018        Appalachian State 356.16
#> 112 2018         Western Kentucky 353.43
#> 113 2018                     Troy 352.36
#> 114 2018               Kent State 351.73
#> 115 2018                     Navy 349.26
#> 116 2018                  Wyoming 348.34
#> 117 2018                     Ohio 346.18
#> 118 2018                     UTEP 345.76
#> 119 2018         Eastern Michigan 344.06
#> 120 2018                  Hawai'i 343.97
#> 121 2018             Old Dominion 333.18
#> 122 2018         Louisiana Monroe 326.66
#> 123 2018                Charlotte 311.00
#> 124 2018         Georgia Southern 303.12
#> 125 2018              Texas State 302.55
#> 126 2018         Coastal Carolina 284.24
#> 127 2018         New Mexico State 249.32
#> 128 2018                     Army 230.75
#> 129 2018                Air Force 186.27
#> 130 2018       Jacksonville State 174.86
#> 131 2018                  Hampton 149.11
#> 132 2018            James Madison 147.11
#> 133 2018   Southeastern Louisiana 135.64
#> 134 2018        Sam Houston State 133.13
#> 135 2018        Sam Houston State 133.13
#> 136 2018              Austin Peay 123.40
#> 137 2018            Alabama State 118.68
#> 138 2018                 Duquesne 113.21
#> 139 2018              Chattanooga 112.15
#> 140 2018           Illinois State 109.44
#> 141 2018        Southern Illinois 108.59
#> 142 2018         Eastern Kentucky 107.58
#> 143 2018              Stony Brook 106.51
#> 144 2018            Montana State  98.98
#> 145 2018                UT Martin  98.34
#> 146 2018                  Liberty  93.99
#> 147 2018                    Idaho  87.35
#> 148 2018           Kennesaw State  86.45
#> 149 2018           Portland State  85.83
#> 150 2018         Northern Arizona  85.03
#> 151 2018                  Samford  84.29
#> 152 2018       North Dakota State  83.68
#> 153 2018                   Towson  83.58
#> 154 2018           Texas Southern  83.08
#> 155 2018          Tennessee State  82.27
#> 156 2018                 Richmond  82.20
#> 157 2018                 Delaware  79.23
#> 158 2018         Youngstown State  70.45
#> 159 2018                Grambling  69.99
#> 160 2018 Southeast Missouri State  69.85
#> 161 2018         Western Illinois  66.84
#> 162 2018                 Nicholls  66.08
#> 163 2018             Rhode Island  62.64
#> 164 2018          Bethune-Cookman  61.90
#> 165 2018                   Furman  59.57
#> 166 2018                 Campbell  58.46
#> 167 2018            Norfolk State  56.44
#> 168 2018           Missouri State  53.43
#> 169 2018              Weber State  51.74
#> 170 2018      Charleston Southern  51.01
#> 171 2018                  Montana  49.16
#> 172 2018             Prairie View  47.10
#> 173 2018        Stephen F. Austin  46.25
#> 174 2018                  McNeese  43.55
#> 175 2018        Abilene Christian  43.10
#> 176 2018       North Carolina A&T  39.98
#> 177 2018              Florida A&M  39.57
#> 178 2018             Murray State  37.82
#> 179 2018                 Cal Poly  37.28
#> 180 2018         Western Carolina  35.54
#> 181 2018             Morgan State  35.38
#> 182 2018     East Tennessee State  35.09
#> 183 2018            Northern Iowa  34.70
#> 184 2018                   Wagner  32.77
#> 185 2018                   Butler  31.53
#> 186 2018                   Albany  31.53
#> 187 2018   North Carolina Central  31.00
#> 188 2018            North Alabama  30.02
#> 189 2018              The Citadel  29.01
#> 190 2018       South Dakota State  28.73
#> 191 2018           William & Mary  28.69
#> 192 2018                 Monmouth  28.13
#> 193 2018           Tennessee Tech  27.77
#> 194 2018                Princeton  27.29
#> 195 2018     South Carolina State  27.01
#> 196 2018          Houston Baptist  26.80
#> 197 2018                   Bryant  24.33
#> 198 2018             North Dakota  23.99
#> 199 2018            Indiana State  23.88
#> 200 2018         Central Arkansas  23.59
#> 201 2018           Incarnate Word  23.03
#> 202 2018             South Dakota  22.43
#> 203 2018                 Southern  22.25
#> 204 2018           Delaware State  21.32
#> 205 2018       Northwestern State  20.95
#> 206 2018              Idaho State  20.50
#> 207 2018         Sacramento State  20.45
#> 208 2018             Alcorn State  20.41
#> 209 2018       Eastern Washington  19.95
#> 210 2018             Sacred Heart  19.62
#> 211 2018                    Maine  18.93
#> 212 2018                   Howard  17.60
#> 213 2018             Jacksonville  17.49
#> 214 2018                  Fordham  15.33
#> 215 2018            Jackson State  14.46
#> 216 2018             Gardner-Webb  14.11
#> 217 2018                   Mercer  13.96
#> 218 2018                    Lamar  13.91
#> 219 2018                  Cornell  13.71
#> 220 2018        Northern Colorado  13.41
#> 221 2018                  Harvard  13.33
#> 222 2018                  Stetson  13.33
#> 223 2018          St Francis (PA)  13.00
#> 224 2018                      VMI  11.90
#> 225 2018            Southern Utah  11.84
#> 226 2018                    Drake  11.10
#> 227 2018               Georgetown  11.10
#> 228 2018            Robert Morris   9.24
#> 229 2018      Arkansas-Pine Bluff   8.98
#> 230 2018                   Dayton   8.88
#> 231 2018               Holy Cross   7.31
#> 232 2018      Central Connecticut   7.09
#> 233 2018         Eastern Illinois   6.75
#> 234 2018                     Elon   6.66
#> 235 2018                   Lehigh   6.66
#> 236 2018                   Marist   4.44
#> 237 2018                     Yale   3.30