Extract Player Stats for a team, year and sport.

Option to select data per season or per game

player_stats(team_id, year, sport, by = "season")

Arguments

team_id
Team ID (form team_mapping)

year
Select year, (example: 2015-2016 season is 2016)

sport
Tradiational Sport Name

by
Stats over season, or stats per game

Examples

player_stats(697,2016,"MBB",by="SEASON")

#>    Jersey                Player  Yr Pos   Ht GP GS      MP FGM FGA   FG% 3FG
#> 1      01              Hogg, DJ  Fr   G  6-9 37  3  673:00  81 212 38.21  46
#> 2      21          Caruso, Alex  Sr   G  6-5 37 37 1067:00 111 221 50.23  28
#> 3      10 Trocha-Morelos, Tonny  So   C 6-10 37 11  644:00 103 223 46.19  30
#> 4      11      Collins, Anthony  Sr   G  6-1 37 37  950:00  52 138 37.68  28
#> 5      03         Gilder, Admon  Fr   G  6-3 37  0  755:00  87 202 43.07  35
#> 6      34          Davis, Tyler  Fr   C 6-10 36 34  820:00 150 229 65.50  NA
#> 7      23         House, Danuel  Sr   G  6-7 36 34 1114:00 185 467 39.61  75
#> 8      42       Miller, Tavario  Jr   F  6-7 36  1  387:00  24  54 44.44  NA
#> 9      12          Jones, Jalen  Sr   F  6-7 33 28  902:00 175 412 42.48  34
#> 10     25         Dobbins, Kyle  Sr   G  6-0 16  0   28:00   4   7 57.14  NA
#> 11     13        Aparicio, Juan  Sr   G  6-4 13  0   22:00   2   6 33.33   1
#> 12     33         Distefano, TJ  Fr   G  6-2 11  0   14:00  NA   3    NA  NA
#> 13     15        Thomas, Elijah  Fr   F  6-9  8  0   79:00  12  19 63.16  NA
#> 14     00        Eubanks, Kobie N/A      6-5  8  0   36:00   4  12 33.33   4
#> 15     22          Byers, Frank  Fr   G  5-9  8  0    9:00   1   3 33.33  NA
#>    3FGA  3FG%  FT FTA   FT% PTS   Avg ORebs DRebs Tot Reb  Avg AST TO STL BLK
#> 1   139 33.09  22  34 64.71 230  6.22    28    78     106 2.86  35 24  20   5
#> 2    76 36.84  51  65 78.46 301  8.14    40    94     134 3.62 185 87  77  14
#> 3    81 37.04  24  48 50.00 260  7.03    61    96     157 4.24  37 27  15  23
#> 4    62 45.16  34  39 87.18 166  4.49     7    47      54 1.46 155 53  38   4
#> 5   101 34.65  51  67 76.12 260  7.03    16    70      86 2.32  50 27  34   5
#> 6    NA    NA 105 168 62.50 405 11.25   106   118     224 6.22  26 58  20  41
#> 7   243 30.86 118 165 71.52 563 15.64    40   132     172 4.78  77 64  19  11
#> 8    NA    NA  26  48 54.17  74  2.06    43    89     132 3.67  12 30  10  12
#> 9   105 32.38 120 171 70.18 504 15.27    64   172     236 7.15  21 53  22   9
#> 10   NA    NA   8  15 53.33  16  1.00     2     1       3 0.19   5  6   1  NA
#> 11    3 33.33  NA   2  0.00   5  0.38    NA    NA           NA   1  1  NA   1
#> 12    1  0.00  NA  NA    NA        NA    NA     2       2 0.18  NA  2  NA  NA
#> 13   NA    NA   6  18 33.33  30  3.75     5    15      20 2.50   6  9  NA   8
#> 14   11 36.36  NA  NA    NA  12  1.50     2     8      10 1.25   3  1   3  NA
#> 15    1  0.00  NA   1  0.00   2  0.25    NA    NA           NA  NA NA   1  NA
#>    Fouls Dbl Dbl Trpl Dbl DQ year team_id
#> 1     44      NA       NA NA 2016     697
#> 2     92      NA       NA NA 2016     697
#> 3     67       1       NA NA 2016     697
#> 4     72      NA       NA NA 2016     697
#> 5     57      NA       NA NA 2016     697
#> 6     95       2       NA NA 2016     697
#> 7     51      NA       NA NA 2016     697
#> 8     62      NA       NA  2 2016     697
#> 9     94       5       NA  4 2016     697
#> 10     2      NA       NA NA 2016     697
#> 11     4      NA       NA NA 2016     697
#> 12     2      NA       NA NA 2016     697
#> 13    23      NA       NA  2 2016     697
#> 14     2      NA       NA NA 2016     697
#> 15     1      NA       NA NA 2016     697

player_stats(697,2016,"WSO",by="game")

#>                   Player      team          opponent  game_date Pos Goals
#> 1           Bates, Emily Texas A&M              Rice 2015-08-21  MF      
#> 2     Crabtree, Brittany Texas A&M              Rice 2015-08-21   F      
#> 3            Hajdu, Kate Texas A&M              Rice 2015-08-21  MF     2
#> 4      Harryman, Ashlynn Texas A&M              Rice 2015-08-21   D      
#> 5        Harvey, Mikaela Texas A&M              Rice 2015-08-21  MF      
#> 6    Malherbe, Stephanie Texas A&M              Rice 2015-08-21   F      
#> 7        Mueller, Karlie Texas A&M              Rice 2015-08-21   D      
#> 8       Paulson, Mckayla Texas A&M              Rice 2015-08-21   D     2
#> 9          Pounds, Haley Texas A&M              Rice 2015-08-21   F      
#> 10       Saucier, Taylor Texas A&M              Rice 2015-08-21  GK      
#> 11         Wright, Grace Texas A&M              Rice 2015-08-21   D      
#> 12      Cousineau, Janae Texas A&M              Rice 2015-08-21          
#> 13       Leonard, Kristi Texas A&M              Rice 2015-08-21          
#> 14            Watt, Ally Texas A&M              Rice 2015-08-21          
#> 15          Bates, Emily Texas A&M  Northwestern St. 2015-08-23  CF      
#> 16           Hajdu, Kate Texas A&M  Northwestern St. 2015-08-23 CDM      
#> 17     Harryman, Ashlynn Texas A&M  Northwestern St. 2015-08-23 LCB      
#> 18       Harvey, Mikaela Texas A&M  Northwestern St. 2015-08-23 CAM     1
#> 19   Malherbe, Stephanie Texas A&M  Northwestern St. 2015-08-23 CAM      
#> 20       Mueller, Karlie Texas A&M  Northwestern St. 2015-08-23  RB      
#> 21      Paulson, Mckayla Texas A&M  Northwestern St. 2015-08-23 RCB      
#> 22         Pounds, Haley Texas A&M  Northwestern St. 2015-08-23  LF     1
#> 23       Saucier, Taylor Texas A&M  Northwestern St. 2015-08-23  GK      
#> 24            Watt, Ally Texas A&M  Northwestern St. 2015-08-23  RF      
#> 25         Wright, Grace Texas A&M  Northwestern St. 2015-08-23  LB      
#> 26      Cousineau, Janae Texas A&M  Northwestern St. 2015-08-23         1
#> 27           Hummel, Mia Texas A&M  Northwestern St. 2015-08-23          
#> 28            Kelly, Amy Texas A&M  Northwestern St. 2015-08-23          
#> 29  Lassandro, Claudette Texas A&M  Northwestern St. 2015-08-23          
#> 30       Leonard, Kristi Texas A&M  Northwestern St. 2015-08-23          
#> 31      Ritchie, Kendall Texas A&M  Northwestern St. 2015-08-23          
#> 32          Bates, Emily Texas A&M         UC Irvine 2015-08-28  CF      
#> 33           Hajdu, Kate Texas A&M         UC Irvine 2015-08-28 CDM     1
#> 34     Harryman, Ashlynn Texas A&M         UC Irvine 2015-08-28 LCB      
#> 35       Harvey, Mikaela Texas A&M         UC Irvine 2015-08-28 CAM     1
#> 36   Malherbe, Stephanie Texas A&M         UC Irvine 2015-08-28 CAM      
#> 37       Mueller, Karlie Texas A&M         UC Irvine 2015-08-28  RB      
#> 38      Paulson, Mckayla Texas A&M         UC Irvine 2015-08-28 RCB      
#> 39         Pounds, Haley Texas A&M         UC Irvine 2015-08-28  LF     1
#> 40       Saucier, Taylor Texas A&M         UC Irvine 2015-08-28  GK      
#>     Assists ShAtt SoG Fouls Red Cards Yellow Cards GC Goal App GGS
#> 1               2   2                                             
#> 2               2                                                 
#> 3               3   2                                             
#> 4                                                                 
#> 5               2                                                 
#> 6               2   1                                             
#> 7               5   1                                             
#> 8               2   2                                             
#> 9     3.00/     6   3                                             
#> 10                                                           1   1
#> 11                                                                
#> 12              2   1                                             
#> 13              1                                1                
#> 14              4   1                                             
#> 15     1.00     4   3                                             
#> 16     1.00     1                                                 
#> 17     1.00                                                       
#> 18              2   2                            1                
#> 19              1                                                 
#> 20              1   1                                             
#> 21              2                                                 
#> 22     1.00     4   2                                             
#> 23                                                           1   1
#> 24              2   1                                             
#> 25                                                                
#> 26              3   2                                             
#> 27                                                           1    
#> 28                                                                
#> 29              1                                                 
#> 30              1   1                                             
#> 31              1                                                 
#> 32     2.00     3   1                                             
#> 33              3   1                                             
#> 34              1                                                 
#> 35              1   1                                             
#> 36              3                                                 
#> 37              3   2                                             
#> 38     1.00     2                                1                
#> 39              5   4                                             
#> 40                                                           1   1
#>     Goalie Min. Plyd GA Saves Shutouts G Wins G Loss DSaves Corners
#> 1                                                                  
#> 2                                                                  
#> 3                                                                  
#> 4                                                                  
#> 5                                                                  
#> 6                                                                  
#> 7                                                                  
#> 8                                                                  
#> 9                                                                  
#> 10             5,400  2  3.00              1/                      
#> 11                                                                 
#> 12                                                                 
#> 13                                                                 
#> 14                                                                 
#> 15                                                                 
#> 16                                                                 
#> 17                                                                 
#> 18                                                                 
#> 19                                                                 
#> 20                                                                 
#> 21                                                                 
#> 22                                                                 
#> 23             2,700  0                    1/                      
#> 24                                                                 
#> 25                                                                 
#> 26                                                                 
#> 27             2,700  0  2.00                                      
#> 28                                                                 
#> 29                                                                 
#> 30                                                                 
#> 31                                                                 
#> 32                                                                 
#> 33                                                                 
#> 34                                                                 
#> 35                                                                 
#> 36                                                                 
#> 37                                                                 
#> 38                                                                 
#> 39                                                                 
#> 40             3,598  0                    1/                      
#>                 location team_id
#> 1                Houston     697
#> 2                Houston     697
#> 3                Houston     697
#> 4                Houston     697
#> 5                Houston     697
#> 6                Houston     697
#> 7                Houston     697
#> 8                Houston     697
#> 9                Houston     697
#> 10               Houston     697
#> 11               Houston     697
#> 12               Houston     697
#> 13               Houston     697
#> 14               Houston     697
#> 15  College Station, Tex     697
#> 16  College Station, Tex     697
#> 17  College Station, Tex     697
#> 18  College Station, Tex     697
#> 19  College Station, Tex     697
#> 20  College Station, Tex     697
#> 21  College Station, Tex     697
#> 22  College Station, Tex     697
#> 23  College Station, Tex     697
#> 24  College Station, Tex     697
#> 25  College Station, Tex     697
#> 26  College Station, Tex     697
#> 27  College Station, Tex     697
#> 28  College Station, Tex     697
#> 29  College Station, Tex     697
#> 30  College Station, Tex     697
#> 31  College Station, Tex     697
#> 32  College Station, Tex     697
#> 33  College Station, Tex     697
#> 34  College Station, Tex     697
#> 35  College Station, Tex     697
#> 36  College Station, Tex     697
#> 37  College Station, Tex     697
#> 38  College Station, Tex     697
#> 39  College Station, Tex     697
#> 40  College Station, Tex     697
#>  [ reached getOption("max.print") -- omitted 356 rows ]